+++ packages/core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @cappa/core
 
+## 0.8.0
+
+### Minor Changes
+
+- eef23fa: Add `connectionTimeout` config option (default: 20s) to prevent indefinite hangs when targets like Storybook are unreachable
+
 ## 0.7.3
 
 ### Patch Changes