+++ packages/plugins/plugin-pages/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @cappa/plugin-pages
 
+## 0.2.0
+
+### Minor Changes
+
+- eef23fa: Add `connectionTimeout` config option (default: 20s) to prevent indefinite hangs when targets like Storybook are unreachable
+
+### Patch Changes
+
+- Updated dependencies [eef23fa]
+  - @cappa/core@0.8.0
+
 ## 0.1.3
 
 ### Patch Changes